+++ ovp_core/serializers/__init__.py
@@ -3,4 +3,5 @@ from .address import *
 from .cause import *
 from .skill import *
 from .lead import *
+from .availability import *
 from .google_address import *